Move DataStreamTestHelper to org.elasticsearch.cluster.metadata to facilitate testing
